Update README.md
A player will need to choose a topic he/she wants to guess and after that, he/she will be given a set of blank spaces that match a random word regarding the chosen topic. Consequently, he/she will have to guess what letters correspond to those blank spaces to reveal the hidden word. If the player’s answer is correct, it will give its definition for better apprehending and then proceed to the next hidden word. However, if the player runs out of tries, the dynamite will explode and the player will lose the game.

These are the parts i'm having trouble creating:
1. I can't setup a set of category button that will redirect to the game.
2. The definitions don't match the guessed word.